Added environment variables
